Use watchdog/sleep mode to wake-up device every 8s
This commit makes the device enter a low power mode at the end of every
loop iteration. Beforehand, the watchdog timer is configured to raise an
interrupt after ~8s. This will cause the controller to wake up once that
interrupt is triggered. The ISR itself is empty, as it is just used for
waking up. Once the device returns from the ISR, the next loop iteration
starts.