Add support for simple `cargo run`
This commit adds a simple cargo runner, a small script, which is invoked
if the user types `cargo run`. This script receives the binary to exe-
cute as its parameter. That runner will automatically download the final
binary to the ATtiny85 microcontroller using any STK500v2-compatible
programmer (of which there are lots of). Currently this uses the Diamex
EXA-Prog.

Of course, different environments need to adjust the runner for their
needs, e.g. a different serial port or a different programmer.

The programming frequency is set rather low, since the CPU might have
been programmed to a very low clock frequency and the ISP frequency must
be at least 8 times lower than the clock frequency itself. This causes
"long" programming times, but since the software itself is likely only
a few hundred bytes in total, this is fine.