Generate a 125Hz hardware PWM on PB0 driving LEDs
This commit configures the 8bit timer 0 to output a PWM signal on OC0A
(PB0) in fast PWM mode. The PWM duty cycle is increased linearly from 0
to 255, which does not feel linear due to the logarithmic curve of per-
ceived brightness.

Note, that the output circuitry needs a FET for driving the LEDs. One
measurement was done at full brightness and shows, that the LEDs draw
roughly 154mA, which is far to high for a single pin. Therefore an ex-
ternal MOSFET (IRFZ14PBF) is used as a low-side switch.