Fade the leds in and out linearly
This commit introduces a fade-out complementing the pre-existing fade-in
with both being linearized for the human eye. This is done using a log-
table tuned for perceived linearity of the brightness.

The commit also makes sure, that the system clock is shut off so that
the long delays between the fadings use minimal energy.