Prevent Timer0 wrap-around in SRF05 measurement
This commit fixes the previously occurring issue of having the timer
count wrapping around on larger distances. Previously, the timer was
clocked with 8MHz/256, with this commit it is clocked slower at
8MHz/128/8. This way, the timer will never overflow and the measurements
are correct, but less accurate. This slower timer has a reading accuracy
of roughly 2.2cm/LSB (`1000000µs/s / (8000000/s /128/8)/58µ`).

With this, a new threshold needs to be selected. Thanks to the now known
resolution, the value is less arbitrary and currently set approximately
to 140cm.