Sleep during SRF05 measurements waking up on edges
This commit makes use of the PCINT edge-detection for sleeping until the
SRF05-sensor activates and deactivates the output signal. This allows
for additional power savings.
This did not work previously, since the SRF05 did output its rising edge
of the output signal exactly at the moment the interrupt handler, that
was triggered by the trigger signal (yes, writing to output pins still
triggers the PCINT ISR), so an edge was lost. This way, the while-loop
did turn into an infinite loop.
Instead, this commit only enables the interrupt when the code is inter-
ested in that interrupt. The system clock is speed up to the necessary
value until no overlapping occurs.