Turn off light only after multiple measurements
Previously, the first measurement, that returned, that nothing was de-
tected, caused the lights to be turned off. In cases, where the causing
measurement was faulty, this causes a power-off/power-on-cycle, which is
unpleasant.
Therefore this commit tries 6 times before turning off. Only after that
many measurements without an object detected, the light turns off.